requirements.txt
google-cloud-secret-manager==2.21.0
omegaconf==2.3.0